pyproject.toml
[project]
name = "crawlconsole"
version = "0.1.0"
description = "Self-hosted crawler with HTML UI using Crawl4AI"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.23",
  "aiohttp>=3.9",
  "aiosqlite>=0.20",
  "beautifulsoup4>=4.12",
  "python-multipart>=0.0.9",
  "robotexclusionrulesparser>=1.7.1",
  "crawl4ai",
  "pydantic-ai-slim[openai]",
  "html2text>=2024.2.26",
  "rank-bm25>=0.2.2",
  "httpx>=0.27",
  "markdownify>=1.2.2",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
